Style: 캠퍼스 핵데이 Java 코딩 컨벤션 적용
참고: https://naver.github.io/hackday-conventions-java/
