+++ src/main/java/cloneproject/Instagram/domain/feed/dto/CommentDTO.java
@@ -5,6 +5,7 @@ import cloneproject.Instagram.domain.member.entity.Member;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.querydsl.core.annotations.QueryProjection;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -14,26 +15,27 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class CommentDTO {
 
-    @JsonIgnore
-    private Long postId;
+	@JsonIgnore
+	private Long postId;
 
-    private Long id;
-    private MemberDTO member;
-    private String content;
-    private LocalDateTime uploadDate;
-    private int commentLikesCount;
-    private boolean commentLikeFlag;
-    private int repliesCount;
+	private Long id;
+	private MemberDTO member;
+	private String content;
+	private LocalDateTime uploadDate;
+	private int commentLikesCount;
+	private boolean commentLikeFlag;
+	private int repliesCount;
 
-    @QueryProjection
-    public CommentDTO(Long postId, Long id, Member member, String content, LocalDateTime uploadDate, int commentLikesCount, boolean commentLikeFlag, int repliesCount) {
-        this.postId = postId;
-        this.id = id;
-        this.member = new MemberDTO(member);
-        this.content = content;
-        this.uploadDate = uploadDate;
-        this.commentLikesCount = commentLikesCount;
-        this.commentLikeFlag = commentLikeFlag;
-        this.repliesCount = repliesCount;
-    }
+	@QueryProjection
+	public CommentDTO(Long postId, Long id, Member member, String content, LocalDateTime uploadDate,
+		int commentLikesCount, boolean commentLikeFlag, int repliesCount) {
+		this.postId = postId;
+		this.id = id;
+		this.member = new MemberDTO(member);
+		this.content = content;
+		this.uploadDate = uploadDate;
+		this.commentLikesCount = commentLikesCount;
+		this.commentLikeFlag = commentLikeFlag;
+		this.repliesCount = repliesCount;
+	}
 }
