+++ src/main/java/cloneproject/Instagram/domain/feed/dto/CommentUploadResponse.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class CommentCreateResponse {
+public class CommentUploadResponse {
 
-    private Long commentId;
+	private Long commentId;
 }
