+++ src/main/java/cloneproject/Instagram/domain/feed/dto/MiniProfilePostDTO.java
@@ -1,6 +1,5 @@
 package cloneproject.Instagram.domain.feed.dto;
 
-
 import com.querydsl.core.annotations.QueryProjection;
 
 import lombok.AccessLevel;
@@ -14,14 +13,13 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MiniProfilePostDTO {
 
-    private Long postId;
-    private String postImageUrl;
+	private Long postId;
+	private String postImageUrl;
 
-    @Builder
-    @QueryProjection
-    public MiniProfilePostDTO(Long postId, String postImageUrl) {
-        this.postId = postId;
-        this.postImageUrl = postImageUrl;
-    }
-
+	@Builder
+	@QueryProjection
+	public MiniProfilePostDTO(Long postId, String postImageUrl) {
+		this.postId = postId;
+		this.postImageUrl = postImageUrl;
+	}
 }
