+++ src/main/java/cloneproject/Instagram/domain/feed/dto/PostDTO.java
@@ -4,6 +4,7 @@ import cloneproject.Instagram.domain.member.dto.MemberDTO;
 import cloneproject.Instagram.domain.member.entity.Member;
 
 import com.querydsl.core.annotations.QueryProjection;
+
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,29 +19,30 @@ import java.util.List;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostDTO {
 
-    private Long postId;
-    private String postContent;
-    private List<PostImageDTO> postImageDTOs = new ArrayList<>();
-    private LocalDateTime postUploadDate;
-    private MemberDTO member;
-    private int postCommentsCount;
-    private int postLikesCount;
-    private boolean postBookmarkFlag;
-    private boolean postLikeFlag;
-    private boolean commentFlag;
-    private String followingMemberUsernameLikedPost;
-    private List<CommentDTO> recentComments = new ArrayList<>();
+	private Long postId;
+	private String postContent;
+	private List<PostImageDTO> postImageDTOs = new ArrayList<>();
+	private LocalDateTime postUploadDate;
+	private MemberDTO member;
+	private int postCommentsCount;
+	private int postLikesCount;
+	private boolean postBookmarkFlag;
+	private boolean postLikeFlag;
+	private boolean commentFlag;
+	private String followingMemberUsernameLikedPost;
+	private List<CommentDTO> recentComments = new ArrayList<>();
 
-    @QueryProjection
-    public PostDTO(Long postId, String postContent, LocalDateTime postUploadDate, Member member, int postCommentsCount, int postLikesCount, boolean postBookmarkFlag, boolean postLikeFlag, boolean commentFlag) {
-        this.postId = postId;
-        this.postContent = postContent;
-        this.postUploadDate = postUploadDate;
-        this.member = new MemberDTO(member);
-        this.postCommentsCount = postCommentsCount;
-        this.postLikesCount = postLikesCount;
-        this.postBookmarkFlag = postBookmarkFlag;
-        this.postLikeFlag = postLikeFlag;
-        this.commentFlag = commentFlag;
-    }
+	@QueryProjection
+	public PostDTO(Long postId, String postContent, LocalDateTime postUploadDate, Member member, int postCommentsCount,
+		int postLikesCount, boolean postBookmarkFlag, boolean postLikeFlag, boolean commentFlag) {
+		this.postId = postId;
+		this.postContent = postContent;
+		this.postUploadDate = postUploadDate;
+		this.member = new MemberDTO(member);
+		this.postCommentsCount = postCommentsCount;
+		this.postLikesCount = postLikesCount;
+		this.postBookmarkFlag = postBookmarkFlag;
+		this.postLikeFlag = postLikeFlag;
+		this.commentFlag = commentFlag;
+	}
 }
