+++ src/main/java/cloneproject/Instagram/domain/feed/dto/PostLikeDTO.java
@@ -1,6 +1,7 @@
 package cloneproject.Instagram.domain.feed.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
+
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,12 +12,12 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostLikeDTO {
 
-    private Long postId;
-    private String username;
+	private Long postId;
+	private String username;
 
-    @QueryProjection
-    public PostLikeDTO(Long postId, String username) {
-        this.postId = postId;
-        this.username = username;
-    }
+	@QueryProjection
+	public PostLikeDTO(Long postId, String username) {
+		this.postId = postId;
+		this.username = username;
+	}
 }
