+++ src/main/java/cloneproject/Instagram/domain/feed/entity/Bookmark.java
@@ -15,22 +15,22 @@ import cloneproject.Instagram.domain.member.entity.Member;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Bookmark {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "bookmark_id")
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "bookmark_id")
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id")
+	private Member member;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "post_id")
+	private Post post;
 
-    @Builder
-    public Bookmark(Member member, Post post) {
-        this.member = member;
-        this.post = post;
-    }
+	@Builder
+	public Bookmark(Member member, Post post) {
+		this.member = member;
+		this.post = post;
+	}
 }
