+++ src/main/java/cloneproject/Instagram/domain/feed/repository/BookmarkRepository.java
@@ -3,11 +3,12 @@ package cloneproject.Instagram.domain.feed.repository;
 import cloneproject.Instagram.domain.feed.entity.Bookmark;
 import cloneproject.Instagram.domain.feed.entity.Post;
 import cloneproject.Instagram.domain.member.entity.Member;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
 
-    Optional<Bookmark> findByMemberAndPost(Member member, Post post);
+	Optional<Bookmark> findByMemberAndPost(Member member, Post post);
 }
