+++ src/main/java/cloneproject/Instagram/domain/feed/repository/querydsl/CommentRepositoryQuerydsl.java
@@ -7,6 +7,7 @@ import cloneproject.Instagram.domain.feed.dto.CommentDTO;
 
 public interface CommentRepositoryQuerydsl {
 
-    Page<CommentDTO> findCommentDtoPage(Long memberId, Long postId, Pageable pageable);
-    Page<CommentDTO> findReplyDtoPage(Long memberId, Long commentId, Pageable pageable);
+	Page<CommentDTO> findCommentDtoPage(Long memberId, Long postId, Pageable pageable);
+
+	Page<CommentDTO> findReplyDtoPage(Long memberId, Long commentId, Pageable pageable);
 }
