+++ src/main/java/cloneproject/Instagram/domain/feed/repository/querydsl/PostLikeRepositoryQuerydsl.java
@@ -7,7 +7,7 @@ import cloneproject.Instagram.domain.member.dto.LikeMembersDTO;
 
 public interface PostLikeRepositoryQuerydsl {
 
-    Page<LikeMembersDTO> findLikeMembersDtoPageByPostIdAndMemberId(Pageable pageable, Long postId, Long memberId);
+	Page<LikeMembersDTO> findLikeMembersDtoPageByPostIdAndMemberId(Pageable pageable, Long postId, Long memberId);
 
-    Page<LikeMembersDTO> findLikeMembersDtoPageByCommentIdAndMemberId(Pageable pageable, Long commentId, Long memberId);
+	Page<LikeMembersDTO> findLikeMembersDtoPageByCommentIdAndMemberId(Pageable pageable, Long commentId, Long memberId);
 }
